Added three new tests that to not end in error states. Updated implementations to fix issues and to implement replacing all occurences of a Term A in a Term B with a Term C.

+++ trunk/source/IcfgInterpreter/src/de/uni_freiburg/informatik/ultimate/plugins/generator/icfginterpreter/NonDeterministicChoice.java
@@ -5,6 +5,8 @@ import java.util.ArrayList;
 import de.uni_freiburg.informatik.ultimate.core.model.preferences.UltimatePreferenceItemGroup;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.variables.IProgramVar;
 import de.uni_freiburg.informatik.ultimate.logic.SMTLIBConstants;
+import de.uni_freiburg.informatik.ultimate.plugins.generator.icfginterpreter.datatypes.BitVector;
+import de.uni_freiburg.informatik.ultimate.plugins.generator.icfginterpreter.datatypes.SMTArray;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.icfginterpreter.interpret.ArrayRestriction;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.icfginterpreter.interpret.BitVectorRestriction;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.icfginterpreter.interpret.BooleanRestriction;
@@ -42,7 +44,7 @@ public interface NonDeterministicChoice {
 		return null;
 	}
 
-	int havocInt(IProgramVar variable, IntegerRestriction values);
+	long havocInt(IProgramVar variable, IntegerRestriction values);
 
 	boolean havocBool(IProgramVar variable, BooleanRestriction values);
